Create a file for ManPageFormatter
